+++ src/WorkflowsRules/Actions/Action.php
@@ -1,0 +1,58 @@
+<?php
+
+namespace Kanvas\Packages\WorkflowsRules\Actions;
+
+use Kanvas\Packages\WorkflowsRules\Contracts\Interfaces\ActionInterfaces;
+use Kanvas\Packages\WorkflowsRules\Contracts\Interfaces\WorkflowsEntityInterfaces;
+
+class Action implements ActionInterfaces
+{
+    protected ?string $message = null;
+
+    protected ?array $data = [];
+
+    protected int $status = 1;
+
+    /**
+     * handle.
+     *
+     * @param  WorkflowsEntityInterfaces $entity
+     * @param  array $params
+     *
+     * @return array
+     */
+    public function handle(WorkflowsEntityInterfaces $entity, array $params = []) : array
+    {
+        return [];
+    }
+
+    /**
+     * getData.
+     *
+     * @return array
+     */
+    public function getData() : ?array
+    {
+        return $this->data;
+    }
+
+    /**
+     * getMessage.
+     *
+     * @return string
+     */
+    public function getMessage() : ?string
+    {
+        return $this->message;
+    }
+
+    /**
+     * getStatus.
+     *
+     * @return bool
+     */
+    public function getStatus() : int
+    {
+        return $this->status;
+    }
+}
