+++ src/WorkflowsRules/Actions/SendToZoho.php
@@ -2,50 +2,13 @@
 
 namespace Kanvas\Packages\WorkflowsRules\Actions;
 
-use Kanvas\Packages\WorkflowsRules\Contracts\Interfaces\ActionInterfaces;
 use Kanvas\Packages\WorkflowsRules\Contracts\Interfaces\WorkflowsEntityInterfaces;
 use Phalcon\Di;
 use Throwable;
 use Zoho\CRM\ZohoClient;
 
-class SendToZoho implements ActionInterfaces
+class SendToZoho extends Action
 {
-    protected ?string $message = null;
-
-    protected ?array $data = [];
-
-    protected int $status = 1;
-
-    /**
-     * getData.
-     *
-     * @return array
-     */
-    public function getData() : ?array
-    {
-        return $this->data;
-    }
-
-    /**
-     * getMessage.
-     *
-     * @return string
-     */
-    public function getMessage() : ?string
-    {
-        return $this->message;
-    }
-
-    /**
-     * getStatus.
-     *
-     * @return bool
-     */
-    public function getStatus() : int
-    {
-        return $this->status;
-    }
-
     /**
      * handle.
      *
